+++ packages/styles/bases/selectbox-input.css.ts
@@ -52,7 +52,6 @@ export const selectBoxInnerBoxMoreStyle = style({
 });
 
 export const selectBoxSelectorStyle = style({
-  padding: THEME.spacing.xs,
   background: THEME.color.white["800"],
   borderRadius: THEME.spacing.xs2,
   boxSizing: "border-box",
@@ -65,22 +64,28 @@ export const selectBoxSelectorOptionStyle = style({
   display: "flex",
   justifyContent: "space-between",
   alignItems: "center",
-  padding: `${THEME.spacing.xs} ${THEME.spacing.xs2}`,
+  padding: THEME.spacing.xs,
   fontSize: THEME.fontSize.sm,
+  lineHeight: THEME.fontSize.xl,
   color: THEME.color.gray["800"],
   boxSizing: "border-box",
   cursor: "pointer",
+  border: "none",
+  backgroundColor: "white",
   "@media": {
     "(any-hover: hover)": {
       ":hover": {
-        color: THEME.color.green["800"],
         background: THEME.color.green["300"],
+      },
+      ":active": {
+        color: THEME.color.white["800"],
+        background: THEME.color.green["800"],
       },
     },
   },
-  ":active": {
-    color: THEME.color.white["800"],
-    background: THEME.color.green["800"],
+  ":disabled": {
+    color: THEME.color.gray["400"],
+    cursor: "not-allowed",
   },
 });