+++ packages/wiz-ui-next/src/components/icons/index.ts
@@ -1,5 +1,6 @@
 import { default as WizIAddCircle } from "./add-circle.vue";
 import { default as WizIAdd } from "./add.vue";
+import { default as WizIAdjustBold } from "./adjust-bold.vue";
 import { default as WizIAdjust } from "./adjust.vue";
 import { default as WizIApprovalFilled } from "./approval-filled.vue";
 import { default as WizIArrowDownward } from "./arrow-downward.vue";
@@ -57,12 +58,14 @@ import { default as WizIHelp } from "./help.vue";
 import { default as WizIHistory } from "./history.vue";
 import { default as WizIHome } from "./home.vue";
 import { default as WizIHourglassTop } from "./hourglass-top.vue";
+import { default as WizIInboxCustomize } from "./inbox-customize.vue";
 import { default as WizIInfoBold } from "./info-bold.vue";
 import { default as WizIInfoFilled } from "./info-filled.vue";
 import { default as WizIInfo } from "./info.vue";
 import { default as WizILocationCity } from "./location-city.vue";
 import { default as WizILocationOn } from "./location-on.vue";
 import { default as WizILogin } from "./login.vue";
+import { default as WizIMail } from "./mail.vue";
 import { default as WizIManageAccounts } from "./manage-accounts.vue";
 import { default as WizIMenuOpen } from "./menu-open.vue";
 import { default as WizIMenu } from "./menu.vue";
@@ -96,6 +99,7 @@ import { default as WizIWarning } from "./warning.vue";
 export type TIcon =
   | typeof WizIAddCircle
   | typeof WizIAdd
+  | typeof WizIAdjustBold
   | typeof WizIAdjust
   | typeof WizIApprovalFilled
   | typeof WizIArrowDownward
@@ -153,12 +157,14 @@ export type TIcon =
   | typeof WizIHistory
   | typeof WizIHome
   | typeof WizIHourglassTop
+  | typeof WizIInboxCustomize
   | typeof WizIInfoBold
   | typeof WizIInfoFilled
   | typeof WizIInfo
   | typeof WizILocationCity
   | typeof WizILocationOn
   | typeof WizILogin
+  | typeof WizIMail
   | typeof WizIManageAccounts
   | typeof WizIMenuOpen
   | typeof WizIMenu
@@ -192,6 +198,7 @@ export type TIcon =
 export {
   WizIAddCircle,
   WizIAdd,
+  WizIAdjustBold,
   WizIAdjust,
   WizIApprovalFilled,
   WizIArrowDownward,
@@ -249,12 +256,14 @@ export {
   WizIHistory,
   WizIHome,
   WizIHourglassTop,
+  WizIInboxCustomize,
   WizIInfoBold,
   WizIInfoFilled,
   WizIInfo,
   WizILocationCity,
   WizILocationOn,
   WizILogin,
+  WizIMail,
   WizIManageAccounts,
   WizIMenuOpen,
   WizIMenu,