+++ packages/wiz-ui-react/src/components/base/snackbar/components/snackbar-provider.tsx
@@ -16,12 +16,13 @@ type SnackbarDataItem = {
 };
 
 const marginPx = (() => {
+  if (!window) return 0;
   const mdRem = getSpacingCss("md");
   if (!mdRem || mdRem === "auto") return 0;
   if (mdRem === "9999px") return 9999;
   return (
     parseFloat(mdRem) *
-    parseFloat(getComputedStyle(document.documentElement).fontSize)
+    parseFloat(window.getComputedStyle(document.documentElement).fontSize)
   );
 })();